Jenkins: subsequent calls to properties wipes out the previous ones
Case EA-7062: fix by collecting properties in an array, and call properties(...)
before FileLoader

In our Jenkinsfile, we erroneously had two calls to properties(...). These calls
ensure, for example, that the "Build Triggers >> Poll SCM" property is set. The
second call is exercised when the branch is master or production, and sets the
default ea-tools branch from which to run ea4-lib.groovy.

Since "last call to properties(...) wins", the end result is:
* master/production were setting the default ea-tools branch, but not "pollSCM"
* non-master/production correctly had the "pollSCM" set, but not default
  ea-tools branch